backend/app/pyproject.toml
[tool.poetry]
name = "oaip"
version = "0.1.0"
description = ""
authors = ["bullbulk <grur.ura@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.7.6"
fastapi = "^0.72.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.3"
PyJWT = "^2.3.0"
pydantic = { extras = ["email"], version = "^1.9.0" }
python-multipart = "^0.0.5"
requests = "^2.27.1"
SQLAlchemy = "^1.4.29"
starlette = "^0.17.1"
tenacity = "^8.0.1"
uvicorn = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
